# [SofaCUDA] No longer use deprecated texture references in TetraTLED