--- a/applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaTetrahedronTLEDForceField.h
+++ b/applications/plugins/SofaCUDA/sofa/gpu/cuda/CudaTetrahedronTLEDForceField.h
@@ -132,6 +132,8 @@ protected:
 
     float* m_device_volume { nullptr };
 
+    float3* m_device_preferredDirection { nullptr };
+
 };
 
 } // namespace cuda