Instead, I use raw gpu pointers.

I measured the performances on the scene `applications\plugins\SofaCUDA\scenes\benchmarks\CudaTetrahedronTLEDForceField_beam10x10x40_gpu.scn` on 50k time steps:

```
before
[INFO]    [BatchGUI] 50000 iterations done in 11.9961 s ( 4168.04 FPS)

after
[INFO]    [BatchGUI] 50000 iterations done in 10.3888 s ( 4812.87 FPS).
```

Note that I did not change anything in the algorithm, except the following check:

```cpp
if (FCrds.y < 0 || FCrds.y >= nbElements)
{
    FCrds.y = 0;
}
```

I had to add this check because `FCrds.y` was often equal to `-1`, which then leads to out-of-bounds vector access. This check was not required before, because out-of-bounds texture returns 0.
I don't know if this behavior is intended or not.

TODO:

- [x] Test that `CudaTetrahedronTLEDForceField` now compiles with CUDA 12.

______________________________________________________

By submitting this pull request, I acknowledge that  
**I have read, understand, and agree [SOFA Developer Certificate of Origin (DCO)](https://github.com/sofa-framework/sofa/blob/master/CONTRIBUTING.md#sofa-developer-certificate-of-origin-dco)**.
______________________________________________________

**Reviewers will merge this pull-request only if**  
- it builds with SUCCESS for all platforms on the CI.
- it does not generate new warnings.
- it does not generate new unit test failures.
- it does not generate new scene test failures.
- it does not break API compatibility.
- it is more than 1 week old (or has fast-merge label).