@fredroy 2023-02-16
It seems to work well here (W11/CUDA12 with 4080)
` 50000 iterations done in 7.96522 s ( 6277.29 FPS)`

Nevertheless, like #3640, I needed to exclude HexaTLED in the cmake (obviously) and commented all the other files using CudaTexture.h (for nothing)